Renamed resources and corresponding files; renamed directory; updated example Terraform files
