+++ examples/network/main.tf
@@ -1,8 +1,8 @@
-provider "enf" {
+provider "xaptum" {
   domain_url = "https://dev.xaptum.io"
 }
 
-resource "enf_network" "network" {
+resource "xaptum_network" "network" {
     name = "TestNetwork 1234"
     description = "Subnet for devices #1-#4 for XYZ company."
 }
