+++ examples/ratelimit/main.tf
@@ -1,10 +1,10 @@
-provider "enf" {
+provider "xaptum" {
   domain_url = "https://dev.xaptum.io"
   username = "" # Need to fill username/password out with your credentials.
   password = ""
 }
 
-resource "enf_endpoint_ratelimit" "test" {
+resource "xaptum_endpoint_ratelimit" "test" {
     endpoint = "fd00:8f80:8000::fd3d:6c94:0"
     inherit = false
     packets_per_second = 500
@@ -13,7 +13,7 @@ resource "enf_endpoint_ratelimit" "test" {
     bytes_burst_size = 500000
 }
 
-resource "enf_network_ratelimit" "test1" {
+resource "xaptum_network_ratelimit" "test1" {
   network = ""
   packets_per_second = 500
   packets_burst_size = 500
@@ -22,7 +22,7 @@ resource "enf_network_ratelimit" "test1" {
   bytes_burst_size = 500000
 }
 
-resource "enf_domain_ratelimit" "test2" {
+resource "xaptum_domain_ratelimit" "test2" {
   domain = ""
   packets_per_second = 1000
   packets_burst_size = 500
@@ -31,12 +31,12 @@ resource "enf_domain_ratelimit" "test2" {
 }
 
 // When the inherit flag is set to true, no rate limit values can be specified. If they are set, an error will be thrown.
-resource "enf_endpoint_ratelimit" "test3" { // Valid resource
+resource "xaptum_endpoint_ratelimit" "test3" { // Valid resource
   endpoint = ""
   inherit = true
 }
 
-resource "enf_endpoint_ratelimit" "test4" { // Invalid resource
+resource "xaptum_endpoint_ratelimit" "test4" { // Invalid resource
   endpoint = ""
   inherit = true
   packets_per_second = 1000
